Fixed remaining bugs. 1.0 release.

+++ DatabaseResources/5 CREATE TABLE WarAttack.sql
@@ -5,7 +5,7 @@ CREATE TABLE WarAttack (
     Stars TINYINT NOT NULL CHECK (Stars >= 0 AND Stars <= 3),
     Destruction_percentage TINYINT NOT NULL CHECK (Destruction_percentage >= 0 AND Destruction_percentage <= 100),
     Order_ TINYINT,
-    Duration TINYINT NOT NULL,
+    Duration int NOT NULL,
     PRIMARY KEY(Player_tag, Clan_war_id, War_attack_number),
     FOREIGN KEY (Player_tag, Clan_war_id) REFERENCES WarParticipation(Player_tag, Clan_war_id) ON DELETE CASCADE
     );
